+++ apps/hightide-native-storybook/src/Internationalization/ExtendTranslations.stories.tsx
@@ -76,7 +76,6 @@ const LocaleSelect = () => {
         {options.map((option) => (
           <Select.Option
             key={option.id}
-            id={option.id}
             value={option.id}
             label={option.label}
           />