+++ packages/hightide/src/components/user-interaction/data/FilterPopUp.tsx
@@ -74,7 +74,7 @@ export const FilterBasePopUp = forwardRef<HTMLDivElement, FilterPopUpBaseProps>(
             triggerProps={{
               'data-name': 'filter-operator-select',
               'className': 'w-fit coloring-text-hover neutral flex-row-1 items-center h-element-sm px-2 py-1 rounded-md hover:cursor-pointer font-bold',
-              'selectedDisplay': (option) => option ? translation(FilterOperatorUtils.getInfo(option.value as FilterOperator).translationKey as Parameters<typeof translation>[0]) : ''
+              'selectedDisplay': (option) => option ? translation(FilterOperatorUtils.getInfo(option.value.value).translationKey as Parameters<typeof translation>[0]) : ''
             }}
             iconAppearance="right"
           >